requirements.txt
Flask==2.0.1
mysql-connector-python==8.3.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
Werkzeug==2.0.1
python-dotenv==0.19.0
email-validator==1.1.3
Flask-Migrate==4.0.5
Pillow==10.0.0
Flask-Mail==0.9.1
SQLAlchemy==1.4.23
Flask-Session==0.4.0   
mysqlclient
pytz==2023.3
WeasyPrint==53.0
Flask-RESTful==0.3.10
requests==2.31.0
gunicorn


